delete.intercept {pls}	R Documentation

Delete intercept from model matrix

Description

A utility function to delete any intercept column from a model matrix, and adjust the "assign" attribute correspondingly. It is used by formula handling functions like mvr and model.matrix.mvr.

Usage

delete.intercept(mm)

Arguments

mm
Model matrix.

Value

A model matrix without intercept column.

Author(s)

Bjørn-Helge Mevik and Ron Wehrens

See Also

mvr, model.matrix.mvr

[Package pls version 2.9-0 Index]