jack.test {pls}	R Documentation

Jackknife approximate t tests of regression coefficients

Description

Performes approximate t tests of regression coefficients based on jackknife variance estimates.

Usage

jack.test(object, ncomp = object$ncomp, use.mean = TRUE)
## S3 method for class 'jacktest'
print(x, P.values = TRUE, ...)

Arguments

object
an mvr object. A cross-validated model fitted with jackknife = TRUE.

ncomp
the number of components to use for estimating the variances

use.mean
logical. If TRUE (default), the mean coefficients are used when estimating the (co)variances; otherwise the coefficients from a model fitted to the entire data set. See var.jack for details.

x
an jacktest object, the result of jack.test.

P.values
logical. Whether to print p values (default).

...
Further arguments sent to the underlying print function printCoefmat.

Details

jack.test uses the variance estimates from var.jack to perform t tests of the regression coefficients. The resulting object has a print method, print.jacktest, which uses printCoefmat for the actual printing.

Value

jack.test returns an object of class "jacktest", with components

coefficients
The estimated regression coefficients

sd
The square root of the jackknife variance estimates

tvalues
The t statistics

df
The ‘degrees of freedom’ used for calculating p values

pvalues
The calculated p values

print.jacktest returns the "jacktest" object (invisibly).

Warning

The jackknife variance estimates are known to be biased (see var.jack). Also, the distribution of the regression coefficient estimates and the jackknife variance estimates are unknown (at least in PLSR/PCR). Consequently, the distribution (and in particular, the degrees of freedom) of the resulting t statistics is unknown. The present code simply assumes a t distribution with m - 1 degrees of freedom, where m is the number of cross-validation segments.

Therefore, the resulting p values should not be used uncritically, and should perhaps be regarded as mere indicator of (non-)significance.

Finally, also keep in mind that as the number of predictor variables increase, the problem of multiple tests increases correspondingly.

Author(s)

Bjørn-Helge Mevik

References

Martens H. and Martens M. (2000) Modified Jack-knife Estimation of Parameter Uncertainty in Bilinear Modelling by Partial Least Squares Regression (PLSR). Food Quality and Preference, 11, 5–16.

See Also

var.jack, mvrCv

Examples

data(oliveoil)
mod <- pcr(sensory ~ chemical, data = oliveoil, validation = "LOO", jackknife = TRUE)
jack.test(mod, ncomp = 2)

[Package pls version 2.5-0 Index]