Adjust for Missing Values

Use missing value information to adjust residuals and predictions. This is the ‘mvr equivalent’ of the naresid.exclude and napredict.exclude functions.

Usage

naExcludeMvr(omit, x, ...)

Arguments

omit

an object produced by an na.action function, typically the "na.action" attribute of the result of na.omit or na.exclude.

x

a three-dimensional array to be adjusted based upon the missing value information in omit.

...

further arguments. Currently not used.

Value

x, padded with NAs along the first dimension (‘rows’).

Details

This is a utility function used to allow predict.mvr and residuals.mvr to compensate for the removal of NAs in the fitting process.

It is called only when the na.action is na.exclude, and pads x with NAs in the correct positions to have the same number of rows as the original data frame.

See also

Author

Bjørn-Helge Mevik and Ron Wehrens