Plot Method for MVR objects

plot.mvr plots predictions, coefficients, scores, loadings, biplots, correlation loadings or validation plots (RMSEP curves, etc.).

# S3 method for mvr
plot(x, plottype = c("prediction", "validation", "coefficients",
                     "scores", "loadings", "biplot", "correlation"), ...)

Arguments

x
an object of class mvr. The fitted model to plot.

plottype
character. What kind of plot to plot.

...
further arguments, sent to the underlying plot functions.

Details

The function is simply a wrapper for the underlying plot functions used to make the selected plots. See predplot.mvr, validationplot, coefplot, scoreplot, loadingplot, biplot.mvr or corrplot for details. Note that all arguments except x and plottype must be named.

Value

plot.mvr returns whatever the underlying plot function returns.

See also

Examples

data(yarn)
nir.pcr <- pcr(density ~ NIR, ncomp = 9, data = yarn, validation = "CV")
if (FALSE) {
plot(nir.pcr, ncomp = 5) # Plot of cross-validated predictions
plot(nir.pcr, "scores") # Score plot
plot(nir.pcr, "loadings", comps = 1:3) # The three first loadings
plot(nir.pcr, "coef", ncomp = 5) # Coefficients
plot(nir.pcr, "val") # RMSEP curves
plot(nir.pcr, "val", val.type = "MSEP", estimate = "CV") # CV MSEP
}