Extract Scores and Loadings from PLSR and PCR Models

These functions extract score and loading matrices from fitted mvr models.

scores(object, ...)
# S3 method for default
scores(object, ...)

loadings(object, ...)
# S3 method for default
loadings(object, ...)

loading.weights(object)

Yscores(object)

Yloadings(object)

Arguments

object
a fitted model to extract from.

...
extra arguments, currently not used.

Details

All functions extract the indicated matrix from the fitted model, and will work with any object having a suitably named component.

The default scores and loadings methods also handle prcomp objects (their scores and loadings components are called x and rotation, resp.), and add an attribute "explvar" with the variance explained by each component, if this is available. (See explvar for details.)

Value

A matrix with scores or loadings.

Note

There is a loadings function in package stats. It simply returns any element named "loadings". See loadings for details. The function can be accessed as stats::loadings(...).

See also

Examples

data(yarn)
plsmod <- plsr(density ~ NIR, 6, data = yarn)
scores(plsmod)

#>           Comp 1       Comp 2       Comp 3       Comp 4      Comp 5
#> 1    0.409179813 -0.008784768  0.523728078 -0.023590598  0.28257621
#> 2    0.290221737  0.086251903  0.282727042  0.129470112 -0.06046771
#> 3    0.335401827 -0.079546117  0.068941588 -0.158106337  0.24942764
#> 4    0.158491452  0.174388578  0.152801468  0.235291233 -0.16995631
#> 5    0.193879472  0.011561311 -0.001745150  0.002704967 -0.17175315
#> 6    0.222557826 -0.171330892 -0.025111808 -0.282910278 -0.21359642
#> 7    0.041285552  0.268429180 -0.159925584  0.225199369  0.20048867
#> 8    0.048482910  0.096640503 -0.103900617  0.139677315 -0.11884022
#> 9    0.101484377 -0.064648475 -0.384088009 -0.076118114  0.43869619
#> 10   0.102290460 -0.259149228 -0.220365644 -0.247923898 -0.14720275
#> 11  -0.100297791  0.299597676  0.042515273 -0.126051034 -0.31191605
#> 12  -0.107259164  0.166769818 -0.009491464  0.130753404 -0.22444482
#> 13  -0.094778901  0.017041622 -0.076235656  0.115579526 -0.07487045
#> 14  -0.076731273 -0.165801118 -0.086620635 -0.012218665 -0.14918846
#> 15  -0.061013351 -0.350839672 -0.120832487 -0.138514150  0.01823246
#> 16  -0.230537785  0.335861951  0.136569922 -0.683816959  0.04495015
#> 17  -0.254876960  0.224228500  0.075408412 -0.171633405  0.23942238
#> 18  -0.270541343  0.090285664  0.141261029  0.081073584  0.12590512
#> 19  -0.274626332 -0.083837842  0.199690644  0.166568223  0.09043754
#> 20  -0.271111813 -0.260331335  0.264987363  0.126705480 -0.03768058
#> 21  -0.255341535 -0.451910791  0.235800476  0.065335843  0.06570659
#> 110  0.133172382  0.054578692 -0.124627601  0.083301894  0.02879014
#> 22   0.134068706 -0.035955155 -0.078724797 -0.026706391 -0.21906668
#> 31  -0.002170868  0.132022347 -0.196235931  0.158751024  0.10454883
#> 41   0.022155628  0.022981746 -0.192941963  0.076277632  0.03129014
#> 51  -0.010006843 -0.118469389 -0.080107295 -0.019172099 -0.25880544
#> 61  -0.098439346  0.124747279 -0.076108930  0.145860618 -0.05596695
#> 71  -0.084938838 -0.054781990 -0.187367726  0.084211702  0.29328394
#>           Comp 6
#> 1    0.193939979
#> 2   -0.053356965
#> 3   -0.103978058
#> 4   -0.132615654
#> 5   -0.015532301
#> 6    0.129509559
#> 7   -0.389401093
#> 8    0.042956978
#> 9   -0.004961549
#> 10   0.024554706
#> 11  -0.325154216
#> 12   0.034059989
#> 13   0.212751948
#> 14   0.138872268
#> 15  -0.189557215
#> 16  -0.044365115
#> 17   0.061696627
#> 18   0.257267306
#> 19   0.229917542
#> 20   0.076976519
#> 21  -0.563830941
#> 110 -0.086860035
#> 22   0.040759248
#> 31  -0.043636342
#> 41   0.158765785
#> 51   0.210824202
#> 61  -0.011601908
#> 71   0.151998735
#> attr(,"class")
#> [1] "scores"
#> attr(,"explvar")
#>     Comp 1     Comp 2     Comp 3     Comp 4     Comp 5     Comp 6 
#> 46.8349936 51.5408548  1.0793049  0.2145479  0.1844248  0.1177235 

loadings(plsmod)[,1:4]

#>               Comp 1        Comp 2        Comp 3        Comp 4
#> NIR1    0.0741483603  0.1604154071 -0.1931935400 -0.0239400227
#> NIR2    0.2387328225  0.3391554985 -0.2386958521 -0.0454774962
#> NIR3    0.4373925427  0.5904104458 -0.2252121027 -0.0653496912
#> NIR4    0.5800264394  0.8583944416 -0.1472653481 -0.0564667700
#> NIR5    0.6016293715  1.0872608096 -0.0556917663 -0.0420557952
#> NIR6    0.5161563261  1.2718581472  0.0030901834 -0.0365405351
#> NIR7    0.3531572365  1.4303571397  0.0253848088 -0.0389396595
#> NIR8    0.1282969872  1.5622222811  0.0218854956 -0.0462387101
#> NIR9   -0.1308499175  1.6794551690  0.0024123580 -0.0509075564
#> NIR10  -0.3958223831  1.8153512286 -0.0277874169 -0.0450939196
#> NIR11  -0.6641502692  1.9799226026 -0.0640975734 -0.0288712888
#> NIR12  -0.9297808414  2.1555547795 -0.1007036726 -0.0053982251
#> NIR13  -1.1557875628  2.3225018929 -0.1355135743  0.0267694682
#> NIR14  -1.3182193914  2.4682377834 -0.1709282279  0.0740056718
#> NIR15  -1.4586292588  2.5954648227 -0.2105910558  0.1363118421
#> NIR16  -1.6409291647  2.7008904066 -0.2566474526  0.2015230837
#> NIR17  -1.8358353455  2.7690680589 -0.2941492923  0.2415733656
#> NIR18  -1.9512747291  2.7831511101 -0.3076503262  0.2307915052
#> NIR19  -1.9405169072  2.7421292849 -0.2938798097  0.1676869203
#> NIR20  -1.8122901437  2.6461203546 -0.2540394105  0.0639891171
#> NIR21  -1.6216674276  2.4843536154 -0.1981967805 -0.0532103032
#> NIR22  -1.4293682568  2.2613046401 -0.1467444408 -0.1440561490
#> NIR23  -1.2424664053  2.0221652004 -0.1115732926 -0.1877755312
#> NIR24  -1.0589354408  1.7812534663 -0.0862383587 -0.1950127610
#> NIR25  -0.9485616833  1.5298457408 -0.0687940799 -0.1901492790
#> NIR26  -0.9379221047  1.2978389885 -0.0577587172 -0.1826022819
#> NIR27  -0.9049070932  1.0634837219 -0.0495838208 -0.1622245893
#> NIR28  -0.7067004492  0.7567267718 -0.0620176600 -0.1033779532
#> NIR29  -0.4370154656  0.4667995366 -0.0753128406 -0.0534623092
#> NIR30  -0.2184258652  0.2631852657 -0.0748881596 -0.0233161791
#> NIR31  -0.0924159328  0.1498612504 -0.0679395260 -0.0058692510
#> NIR32  -0.0584859186  0.1066876155 -0.0632397482  0.0000780767
#> NIR33  -0.0813919962  0.1061429591 -0.0639740763 -0.0010424222
#> NIR34  -0.1175962532  0.1242608780 -0.0680683629 -0.0045039730
#> NIR35  -0.1551773107  0.1296698501 -0.0730965451 -0.0070978793
#> NIR36  -0.2062580375  0.0917541514 -0.0800994911 -0.0072361058
#> NIR37  -0.2669383345  0.0376422459 -0.0890970745 -0.0061900362
#> NIR38  -0.3235119918  0.0258336511 -0.0978405863 -0.0080110312
#> NIR39  -0.3656856740  0.0340655611 -0.1057635264 -0.0111844540
#> NIR40  -0.3828480019  0.0092651923 -0.1121558897 -0.0096967903
#> NIR41  -0.3655078661 -0.0269339459 -0.1143819312 -0.0054619918
#> NIR42  -0.3340756750 -0.0187643145 -0.1109950578 -0.0044026653
#> NIR43  -0.3292284126  0.0202365891 -0.1065412445 -0.0070858043
#> NIR44  -0.3518787056  0.0257932474 -0.1084044654 -0.0092229375
#> NIR45  -0.3751723373 -0.0138013361 -0.1154863536 -0.0080990939
#> NIR46  -0.3837466462 -0.0207784619 -0.1201560274 -0.0089665622
#> NIR47  -0.3768316068  0.0530633452 -0.1175670701 -0.0150364721
#> NIR48  -0.3862588164  0.1213234977 -0.1140580650 -0.0214897473
#> NIR49  -0.4296882403  0.1270471916 -0.1213122141 -0.0239663046
#> NIR50  -0.4827664642  0.1108331143 -0.1358097061 -0.0255565906
#> NIR51  -0.5110094095  0.0434932901 -0.1512012542 -0.0215705780
#> NIR52  -0.4666649289 -0.1790387553 -0.1707725071  0.0025042688
#> NIR53  -0.3609060751 -0.4275437002 -0.1842444028  0.0233456325
#> NIR54  -0.2643699746 -0.5115207306 -0.1803706589  0.0337808413
#> NIR55  -0.1851850784 -0.3865232113 -0.1595724044  0.0319140900
#> NIR56  -0.1309215643 -0.1171838176 -0.1285883300  0.0082919386
#> NIR57  -0.1135732514  0.1253834533 -0.0989369680 -0.0245761107
#> NIR58  -0.0550565771  0.1435692730 -0.0889339542 -0.0254046520
#> NIR59   0.0532298501  0.0097420806 -0.0852074221 -0.0037923550
#> NIR60   0.0919246986 -0.1178337069 -0.0819773469  0.0231981497
#> NIR61   0.0576092058 -0.2128517989 -0.0924097756  0.0460810889
#> NIR62   0.0706542375 -0.2292484778 -0.1017709160  0.0559662050
#> NIR63   0.1458904566 -0.1676330944 -0.0951809097  0.0546122408
#> NIR64   0.1591759067 -0.1151393407 -0.0829709008  0.0464889991
#> NIR65   0.0881304572 -0.0561309547 -0.0746763225  0.0266164562
#> NIR66   0.0802605878  0.1183990348 -0.0737014966 -0.0016686251
#> NIR67   0.1676941546  0.2352252690 -0.0629852024 -0.0146908092
#> NIR68   0.2914118046  0.4964959669 -0.0701288784 -0.0270924198
#> NIR69   0.7145301820  0.8836668744 -0.0488286425 -0.0329452680
#> NIR70   1.1197109724  0.7994978353  0.0261514199  0.0240618453
#> NIR71   1.2760432522  0.6108850235  0.0627945281  0.0490677142
#> NIR72   1.2979061154  0.4829525854  0.0681895695  0.0600822686
#> NIR73   1.2722659556  0.4186312349  0.0645261777  0.0650601575
#> NIR74   1.2364421674  0.3868834966  0.0594106933  0.0661678452
#> NIR75   1.2027098200  0.3676388715  0.0536092810  0.0656260556
#> NIR76   1.1782765615  0.3541143655  0.0477320725  0.0644931191
#> NIR77   1.1682106095  0.3458326206  0.0418194277  0.0634709552
#> NIR78   1.1728572858  0.3429901270  0.0351188080  0.0627498082
#> NIR79   1.1895327523  0.3444056128  0.0266353684  0.0626439553
#> NIR80   1.2163438735  0.3491461694  0.0176778265  0.0622267584
#> NIR81   1.2494546517  0.3575054793  0.0085026234  0.0608481666
#> NIR82   1.2805270999  0.3703477290 -0.0047175161  0.0597281139
#> NIR83   1.3016320242  0.3869447664 -0.0242754376  0.0586209887
#> NIR84   1.3107389228  0.4031858446 -0.0502231566  0.0565197620
#> NIR85   1.3079045726  0.4131715086 -0.0797171438  0.0543905659
#> NIR86   1.2899825968  0.4139734731 -0.1119123500  0.0525597902
#> NIR87   1.2629606380  0.4078159803 -0.1475610447  0.0491859607
#> NIR88   1.2307534759  0.4027644789 -0.1814687413  0.0443981647
#> NIR89   1.1884589987  0.4076642651 -0.2130351550  0.0425927798
#> NIR90   1.1450458271  0.4209518473 -0.2397087577  0.0368291926
#> NIR91   1.1162032626  0.4352226013 -0.2507985882  0.0231996229
#> NIR92   1.0993048067  0.4499443517 -0.2552473696  0.0107458315
#> NIR93   1.0852249069  0.4677180756 -0.2610998183  0.0013992605
#> NIR94   1.0794296946  0.4863708304 -0.2667764115 -0.0035578311
#> NIR95   1.0852975483  0.5071963374 -0.2763927659 -0.0042091854
#> NIR96   1.1051597036  0.5310181721 -0.2813566617 -0.0025486447
#> NIR97   1.1350983055  0.5494108839 -0.2802487833  0.0011990524
#> NIR98   1.1601384495  0.5581229086 -0.2769773108  0.0049763595
#> NIR99   1.1811114678  0.5625195900 -0.2630693466  0.0055265150
#> NIR100  1.2049447034  0.5651882821 -0.2409276334  0.0040286629
#> NIR101  1.2259618108  0.5702573022 -0.2213017746  0.0059794876
#> NIR102  1.2360452416  0.5774796031 -0.2050465578  0.0104852502
#> NIR103  1.2360032290  0.5832588188 -0.1903738750  0.0121960959
#> NIR104  1.2324224566  0.5901899189 -0.1749124584  0.0130492582
#> NIR105  1.2250098767  0.5949735131 -0.1573322904  0.0109678115
#> NIR106  1.2104000305  0.5980553652 -0.1441594374  0.0067857850
#> NIR107  1.1917605531  0.6031315120 -0.1361594778  0.0055333531
#> NIR108  1.1709369841  0.6084708929 -0.1283820178  0.0040707295
#> NIR109  1.1495867661  0.6155138484 -0.1204532284  0.0014675270
#> NIR110  1.1327489808  0.6254308844 -0.1119616430 -0.0003332830
#> NIR111  1.1183918017  0.6367456232 -0.1048162072 -0.0014269434
#> NIR112  1.1031596503  0.6500007959 -0.1009880354 -0.0037184872
#> NIR113  1.0880071593  0.6664239861 -0.0956513446 -0.0095545342
#> NIR114  1.0720944521  0.6870080470 -0.0835047580 -0.0152097839
#> NIR115  1.0530017483  0.7122344562 -0.0693076913 -0.0189695561
#> NIR116  1.0365569107  0.7386729849 -0.0609310277 -0.0226284375
#> NIR117  1.0291400691  0.7668856216 -0.0537699186 -0.0248242107
#> NIR118  1.0265015006  0.7978326777 -0.0442845726 -0.0268873576
#> NIR119  1.0228441575  0.8265587815 -0.0337190928 -0.0299187442
#> NIR120  1.0169899581  0.8518186122 -0.0234448259 -0.0326886487
#> NIR121  1.0101424562  0.8747643381 -0.0120641182 -0.0354612241
#> NIR122  1.0015719063  0.8942954505  0.0002400852 -0.0367963344
#> NIR123  0.9919797065  0.9123160740  0.0096984410 -0.0380556197
#> NIR124  0.9848908253  0.9357652817  0.0182136810 -0.0400215988
#> NIR125  0.9799594780  0.9664075538  0.0300942638 -0.0413923158
#> NIR126  0.9752290905  0.9946815548  0.0433808313 -0.0426625339
#> NIR127  0.9722961714  1.0113283229  0.0544406697 -0.0442702908
#> NIR128  0.9652944890  1.0164582308  0.0675713655 -0.0439304508
#> NIR129  0.9348034557  1.0117697554  0.0827688840 -0.0414880834
#> NIR130  0.8761566260  0.9972974911  0.0918841442 -0.0401893283
#> NIR131  0.8118188232  0.9725477905  0.0942121915 -0.0410505765
#> NIR132  0.7586375481  0.9347204442  0.0934507062 -0.0415465255
#> NIR133  0.7189239134  0.8868321616  0.0913273301 -0.0400584896
#> NIR134  0.6877077329  0.8391874392  0.0887279360 -0.0383947758
#> NIR135  0.6548386782  0.7975166173  0.0855534695 -0.0379615514
#> NIR136  0.6121697285  0.7569403593  0.0814493050 -0.0381327923
#> NIR137  0.5664704616  0.7134544248  0.0759528961 -0.0379173657
#> NIR138  0.5332804660  0.6700900694  0.0700202931 -0.0371677604
#> NIR139  0.5143222670  0.6310860426  0.0654797388 -0.0356870891
#> NIR140  0.4987348023  0.5905182775  0.0615364431 -0.0330797793
#> NIR141  0.4790169653  0.5348961331  0.0576172724 -0.0286358321
#> NIR142  0.4474411936  0.4552012837  0.0562666709 -0.0225397941
#> NIR143  0.4103730318  0.3614305324  0.0573038040 -0.0153389444
#> NIR144  0.4051669962  0.2866240781  0.0537398343 -0.0092454594
#> NIR145  0.4351727623  0.2533475173  0.0411863778 -0.0060147463
#> NIR146  0.4437556826  0.2504474162  0.0265871893 -0.0048060686
#> NIR147  0.3840762024  0.2510771915  0.0168145900 -0.0054387271
#> NIR148  0.2571887612  0.2332363895  0.0072226928 -0.0087017142
#> NIR149  0.1106163347  0.1957370090 -0.0066785005 -0.0130956783
#> NIR150  0.0113648788  0.1614874504 -0.0208938256 -0.0160983895
#> NIR151 -0.0066668777  0.1470177554 -0.0299737865 -0.0161997634
#> NIR152  0.0346515085  0.1327483749 -0.0329288032 -0.0110577242
#> NIR153  0.0826470769  0.1108788117 -0.0328010548 -0.0034414049
#> NIR154  0.1075051771  0.1065313146 -0.0322074256  0.0009384429
#> NIR155  0.1307688889  0.1163900217 -0.0307011698  0.0027828269
#> NIR156  0.1722122504  0.1159253610 -0.0269011481  0.0060288148
#> NIR157  0.2139435483  0.1133098608 -0.0211826690  0.0092531414
#> NIR158  0.2319243598  0.1246168633 -0.0162091518  0.0098539356
#> NIR159  0.2248338251  0.1421333158 -0.0140893262  0.0081767455
#> NIR160  0.2015465243  0.1537429234 -0.0151131571  0.0060422522
#> NIR161  0.1780073360  0.1560700979 -0.0180129204  0.0045653412
#> NIR162  0.1798980891  0.1471134053 -0.0192246080  0.0051709609
#> NIR163  0.1970894986  0.1302457883 -0.0164654726  0.0077803541
#> NIR164  0.1901602746  0.1146441931 -0.0150746996  0.0093035030
#> NIR165  0.1580162558  0.1015625474 -0.0185234823  0.0087479378
#> NIR166  0.1240637703  0.0892205245 -0.0238995273  0.0077656328
#> NIR167  0.0984232477  0.0779841991 -0.0286190374  0.0073402975
#> NIR168  0.0788213871  0.0683757962 -0.0320040419  0.0071685143
#> NIR169  0.0625630864  0.0600070618 -0.0346245459  0.0070840343
#> NIR170  0.0488311005  0.0525230126 -0.0371236433  0.0071821621
#> NIR171  0.0371947578  0.0458917497 -0.0392955388  0.0072512342
#> NIR172  0.0275603205  0.0402342918 -0.0409770107  0.0072223121
#> NIR173  0.0199832392  0.0356166728 -0.0424560208  0.0072572277
#> NIR174  0.0137596346  0.0316739002 -0.0437455610  0.0072612663
#> NIR175  0.0081161658  0.0279832641 -0.0446374527  0.0071930647
#> NIR176  0.0029669966  0.0245922654 -0.0455172829  0.0072286247
#> NIR177 -0.0018990375  0.0216241453 -0.0464105549  0.0072750624
#> NIR178 -0.0067814583  0.0189249316 -0.0470887722  0.0071676738
#> NIR179 -0.0112984817  0.0164751411 -0.0477693683  0.0070646414
#> NIR180 -0.0151238620  0.0144599468 -0.0485365804  0.0070736986
#> NIR181 -0.0184056481  0.0129592571 -0.0490751495  0.0069523222
#> NIR182 -0.0209353467  0.0119407992 -0.0494221410  0.0067901932
#> NIR183 -0.0223940360  0.0112566879 -0.0498905851  0.0068041273
#> NIR184 -0.0232850200  0.0103166261 -0.0501505088  0.0068596993
#> NIR185 -0.0241272148  0.0088667958 -0.0501845992  0.0068951026
#> NIR186 -0.0248236641  0.0074102285 -0.0503885414  0.0070204378
#> NIR187 -0.0254239260  0.0065131196 -0.0506081543  0.0070799605
#> NIR188 -0.0259874272  0.0061206308 -0.0506695021  0.0069933193
#> NIR189 -0.0259406523  0.0059256217 -0.0507455594  0.0070026657
#> NIR190 -0.0250813008  0.0056009584 -0.0508536350  0.0071148886
#> NIR191 -0.0238391698  0.0048874124 -0.0506515634  0.0072102081
#> NIR192 -0.0223003488  0.0037328461 -0.0504241492  0.0073691003
#> NIR193 -0.0204547704  0.0024218285 -0.0504055689  0.0076604862
#> NIR194 -0.0184411153  0.0015051952 -0.0502251418  0.0077937735
#> NIR195 -0.0159467586  0.0016215573 -0.0498839476  0.0077652372
#> NIR196 -0.0128216664  0.0028007451 -0.0495912416  0.0078218284
#> NIR197 -0.0097987629  0.0041323344 -0.0491839477  0.0078839635
#> NIR198 -0.0078503790  0.0049943107 -0.0485900013  0.0078694751
#> NIR199 -0.0082231966  0.0054376321 -0.0482400902  0.0078615822
#> NIR200 -0.0128120169  0.0051594984 -0.0481244129  0.0077100459
#> NIR201 -0.0223273068  0.0035811692 -0.0482499203  0.0072878720
#> NIR202 -0.0356155387  0.0009081289 -0.0490246699  0.0067857295
#> NIR203 -0.0513385376 -0.0022832068 -0.0504229752  0.0062337821
#> NIR204 -0.0684002884 -0.0053948318 -0.0520426578  0.0053746501
#> NIR205 -0.0853415283 -0.0077495132 -0.0538179671  0.0044018959
#> NIR206 -0.1011319652 -0.0097667004 -0.0558055016  0.0035584450
#> NIR207 -0.1159741582 -0.0122586514 -0.0576122351  0.0027525435
#> NIR208 -0.1303445330 -0.0147049174 -0.0591995056  0.0018987575
#> NIR209 -0.1443121040 -0.0162179075 -0.0608911685  0.0010433832
#> NIR210 -0.1567443743 -0.0170136021 -0.0627793788  0.0001465063
#> NIR211 -0.1647480889 -0.0181215986 -0.0647507545 -0.0005210089
#> NIR212 -0.1667215678 -0.0209188730 -0.0666276888 -0.0005537706
#> NIR213 -0.1646977505 -0.0267608534 -0.0680246236 -0.0001389930
#> NIR214 -0.1621338931 -0.0350945865 -0.0687853059  0.0003768218
#> NIR215 -0.1605378441 -0.0442242321 -0.0695860132  0.0009917884
#> NIR216 -0.1597413293 -0.0541920381 -0.0706411871  0.0016693251
#> NIR217 -0.1597712096 -0.0674063103 -0.0717408387  0.0024985385
#> NIR218 -0.1608883803 -0.0879102517 -0.0734556306  0.0038912185
#> NIR219 -0.1599551560 -0.1180192224 -0.0764294654  0.0062882065
#> NIR220 -0.1473007387 -0.1534908220 -0.0801301935  0.0099623958
#> NIR221 -0.1135440080 -0.1857552294 -0.0826058547  0.0146828508
#> NIR222 -0.0635483147 -0.2089041585 -0.0814125342  0.0195733306
#> NIR223 -0.0193000077 -0.2225269521 -0.0758261112  0.0234460801
#> NIR224 -0.0009643496 -0.2290002960 -0.0696935203  0.0254860615
#> NIR225 -0.0078537121 -0.2289346359 -0.0669455873  0.0258782574
#> NIR226 -0.0314396738 -0.2192345323 -0.0664166452  0.0245055310
#> NIR227 -0.0746062810 -0.1904418272 -0.0665553003  0.0201636615
#> NIR228 -0.1384370840 -0.1335862150 -0.0687812005  0.0119226287
#> NIR229 -0.2027260946 -0.0585055277 -0.0731743292  0.0013659246
#> NIR230 -0.2345692463  0.0099389218 -0.0764917922 -0.0074665867
#> NIR231 -0.2156764462  0.0521206307 -0.0756666618 -0.0110429953
#> NIR232 -0.1589788341  0.0629316225 -0.0701257717 -0.0088414625
#> NIR233 -0.0966601222  0.0563862266 -0.0621933251 -0.0042594988
#> NIR234 -0.0544227609  0.0492031253 -0.0548733874 -0.0004830207
#> NIR235 -0.0379193956  0.0441461452 -0.0501990811  0.0018481989
#> NIR236 -0.0351637365  0.0395277809 -0.0486611132  0.0029272201
#> NIR237 -0.0309299796  0.0367597063 -0.0487458384  0.0033464960
#> NIR238 -0.0206349200  0.0358712165 -0.0487117436  0.0038382025
#> NIR239 -0.0081870113  0.0342664347 -0.0479040757  0.0045446013
#> NIR240  0.0039302554  0.0287925436 -0.0468304510  0.0056712425
#> NIR241  0.0154234020  0.0185847107 -0.0462102441  0.0074317823
#> NIR242  0.0261521657  0.0074944546 -0.0459737604  0.0092952310
#> NIR243  0.0366519530 -0.0007405944 -0.0454607566  0.0106838377
#> NIR244  0.0440667839 -0.0071709271 -0.0445049414  0.0117562197
#> NIR245  0.0423435161 -0.0129052352 -0.0438383238  0.0123864232
#> NIR246  0.0297235127 -0.0150712098 -0.0438921566  0.0120934853
#> NIR247  0.0101938917 -0.0109479085 -0.0447189657  0.0107942170
#> NIR248 -0.0114272803 -0.0016331463 -0.0461979048  0.0087970141
#> NIR249 -0.0278196523  0.0106943336 -0.0479309321  0.0065676629
#> NIR250 -0.0311715143  0.0225000405 -0.0488436381  0.0051122801
#> NIR251 -0.0245404830  0.0291105375 -0.0483696921  0.0050415967
#> NIR252 -0.0173582886  0.0293506680 -0.0472047959  0.0056731914
#> NIR253 -0.0128723589  0.0246834286 -0.0462643019  0.0063575362
#> NIR254 -0.0098540933  0.0173698686 -0.0460327185  0.0071614301
#> NIR255 -0.0086194752  0.0096943152 -0.0462447578  0.0080041882
#> NIR256 -0.0096283730  0.0030895455 -0.0466673484  0.0086364125
#> NIR257 -0.0121172943 -0.0020875704 -0.0475161814  0.0091646711
#> NIR258 -0.0158198559 -0.0061318350 -0.0485284597  0.0094696282
#> NIR259 -0.0204671140 -0.0090405341 -0.0493704993  0.0094091086
#> NIR260 -0.0249258210 -0.0105073281 -0.0501464066  0.0092315208
#> NIR261 -0.0282609035 -0.0113699661 -0.0509220595  0.0091297697
#> NIR262 -0.0307726467 -0.0127151557 -0.0513972245  0.0090514870
#> NIR263 -0.0328082146 -0.0140819410 -0.0517306968  0.0090248057
#> NIR264 -0.0342628659 -0.0147274759 -0.0521745043  0.0090947436
#> NIR265 -0.0353761638 -0.0148710131 -0.0524369547  0.0091074984
#> NIR266 -0.0361260390 -0.0149963730 -0.0524656169  0.0090149688
#> NIR267 -0.0364028401 -0.0152672534 -0.0525303511  0.0090100077
#> NIR268 -0.0364405491 -0.0156934585 -0.0526432200  0.0090917580