Extract Scores and Loadings from PLSR and PCR Models

These functions extract score and loading matrices from fitted mvr models.

Usage

loadings(object, ...)

# Default S3 method
loadings(object, ...)

scores(object, ...)

# Default S3 method
scores(object, estimate, newdata, ...)

Yscores(object)

loading.weights(object)

Yloadings(object)

Arguments

object

a fitted model to extract from.

...

extra arguments, currently not used.

estimate

optional character vector ("train", "CV", "test") used by scores to select the desired estimate.

newdata

optional data frame passed to scores when estimate = "test".

Value

A matrix with scores or loadings.

Details

All functions extract the indicated matrix from the fitted model, and will work with any object having a suitably named component.

The default scores and loadings methods also handle prcomp objects (their scores and loadings components are called x and rotation, resp.), and add an attribute "explvar" with the variance explained by each component, if this is available. (See explvar for details.)

Note

There is a loadings function in package stats. It simply returns any element named "loadings". See loadings for details. The function can be accessed as stats::loadings(...).

See also

Author

Ron Wehrens and Bjørn-Helge Mevik

Examples

data(yarn)
plsmod <- plsr(density ~ NIR, 6, data = yarn)
scores(plsmod)
#>          Comp 1      Comp 2       Comp 3       Comp 4      Comp 5       Comp 6
#> 1    4.57990419 -0.09140373  0.869242045 -0.017624270  0.18502203  0.099904242
#> 2    3.24841966  0.89743358  0.469247768  0.096725666 -0.03959236 -0.027485757
#> 3    3.75411539 -0.82766123  0.114423744 -0.118119468  0.16331738 -0.053562185
#> 4    1.77397722  1.81447783  0.253607676  0.175783437 -0.11128205 -0.068314261
#> 5    2.17007140  0.12029310 -0.002896461  0.002020851 -0.11245856 -0.008001149
#> 6    2.49106503 -1.78266322 -0.041678574 -0.211359090 -0.13985623  0.066714219
#> 7    0.46210460  2.79295125 -0.265431714  0.168243919  0.13127368 -0.200592065
#> 8    0.54266383  1.00552486 -0.172445948  0.104351353 -0.07781285  0.022128415
#> 9    1.13590335 -0.67265429 -0.637478608 -0.056866988  0.28724448 -0.002555841
#> 10   1.14492576 -2.69639522 -0.365745299 -0.185221159 -0.09638374  0.012648858
#> 11  -1.12262204  3.11725314  0.070563455 -0.094171311 -0.20423283 -0.167496591
#> 12  -1.20053991  1.73520617 -0.015753174  0.097684399 -0.14695942  0.017545312
#> 13  -1.06084971  0.17731462 -0.126529855  0.086348166 -0.04902282  0.109594846
#> 14  -0.85884461 -1.72512705 -0.143766013 -0.009128428 -0.09768392  0.071537229
#> 15  -0.68291566 -3.65041572 -0.200548113 -0.103482366  0.01193804 -0.097646550
#> 16  -2.58038381  3.49457557  0.226667852 -0.510871968  0.02943195 -0.022853788
#> 17  -2.85280950  2.33305213  0.125156861 -0.128225389  0.15676625  0.031781764
#> 18  -3.02813920  0.93940405  0.234453776  0.060569164  0.08243872  0.132526028
#> 19  -3.07386202 -0.87231577  0.331430586  0.124441248  0.05921566  0.118437352
#> 20  -3.03452441 -2.70869481  0.439804866  0.094660241 -0.02467206  0.039652890
#> 21  -2.85800943 -4.70204024  0.391362802  0.048811675  0.04302261 -0.290446059
#> 110  1.49058368  0.56788023 -0.206846940  0.062233909  0.01885088 -0.044744183
#> 22   1.50061613 -0.37410610 -0.130661131 -0.019952044 -0.14343798  0.020996299
#> 31  -0.02429829  1.37366578 -0.325696729  0.118601107  0.06845529 -0.022478376
#> 41   0.24798549  0.23912041 -0.320229664  0.056986162  0.02048780  0.081784970
#> 51  -0.11200548 -1.23265000 -0.132955691 -0.014323259 -0.16945767  0.108601806
#> 61  -1.10182066  1.29797017 -0.126319525  0.108970829 -0.03664540 -0.005976487
#> 71  -0.95071098 -0.56999551 -0.310977991  0.062913616  0.19203311  0.078299062
#> attr(,"class")
#> [1] "scores"
#> attr(,"explvar")
#>     Comp 1     Comp 2     Comp 3     Comp 4     Comp 5     Comp 6 
#> 46.8349936 51.5408548  1.0793049  0.2145479  0.1844248  0.1177235 
loadings(plsmod)[,1:4]
#>               Comp 1        Comp 2       Comp 3        Comp 4
#> NIR1    6.624595e-03  1.541745e-02 -0.116401274 -0.0320444153
#> NIR2    2.132897e-02  3.259607e-02 -0.143816927 -0.0608729488
#> NIR3    3.907772e-02  5.674406e-02 -0.135692817 -0.0874724587
#> NIR4    5.182098e-02  8.249987e-02 -0.088729023 -0.0755824107
#> NIR5    5.375104e-02  1.044961e-01 -0.033554914 -0.0562929027
#> NIR6    4.611467e-02  1.222377e-01  0.001861870 -0.0489105670
#> NIR7    3.155193e-02  1.374709e-01  0.015294632 -0.0521218645
#> NIR8    1.146237e-02  1.501444e-01  0.013186256 -0.0618918557
#> NIR9   -1.169045e-02  1.614116e-01  0.001453473 -0.0681412420
#> NIR10  -3.536374e-02  1.744725e-01 -0.016742230 -0.0603595204
#> NIR11  -5.933681e-02  1.902894e-01 -0.038619507 -0.0386450581
#> NIR12  -8.306889e-02  2.071693e-01 -0.060675092 -0.0072256811
#> NIR13  -1.032609e-01  2.232145e-01 -0.081648448  0.0358317103
#> NIR14  -1.177729e-01  2.372211e-01 -0.102986174  0.0990587399
#> NIR15  -1.303175e-01  2.494489e-01 -0.126883472  0.1824573576
#> NIR16  -1.466046e-01  2.595813e-01 -0.154632968  0.2697444974
#> NIR17  -1.640180e-01  2.661338e-01 -0.177228247  0.3233529623
#> NIR18  -1.743316e-01  2.674873e-01 -0.185362771  0.3089211292
#> NIR19  -1.733705e-01  2.635447e-01 -0.177065880  0.2244538105
#> NIR20  -1.619144e-01  2.543173e-01 -0.153061593  0.0856512907
#> NIR21  -1.448837e-01  2.387700e-01 -0.119415782 -0.0712235354
#> NIR22  -1.277032e-01  2.173329e-01 -0.088415171 -0.1928233371
#> NIR23  -1.110050e-01  1.943493e-01 -0.067224160 -0.2513429993
#> NIR24  -9.460788e-02  1.711954e-01 -0.051959578 -0.2610302415
#> NIR25  -8.474681e-02  1.470327e-01 -0.041449205 -0.2545203297
#> NIR26  -8.379625e-02  1.247347e-01 -0.034800275 -0.2444184550
#> NIR27  -8.084661e-02  1.022109e-01 -0.029874808 -0.2171423219
#> NIR28  -6.313834e-02  7.272864e-02 -0.037366335 -0.1383743913
#> NIR29  -3.904403e-02  4.486388e-02 -0.045376831 -0.0715608527
#> NIR30  -1.951470e-02  2.529461e-02 -0.045120956 -0.0312093825
#> NIR31  -8.256665e-03  1.440309e-02 -0.040934326 -0.0078561628
#> NIR32  -5.225275e-03  1.025369e-02 -0.038102657  0.0001045079
#> NIR33  -7.271759e-03  1.020135e-02 -0.038545098 -0.0013953124
#> NIR34  -1.050634e-02  1.194265e-02 -0.041011952 -0.0060286985
#> NIR35  -1.386392e-02  1.246251e-02 -0.044041488 -0.0095007175
#> NIR36  -1.842760e-02  8.818447e-03 -0.048260842 -0.0096857377
#> NIR37  -2.384892e-02  3.617778e-03 -0.053681986 -0.0082855432
#> NIR38  -2.890335e-02  2.482860e-03 -0.058950050 -0.0107229978
#> NIR39  -3.267125e-02  3.274024e-03 -0.063723711 -0.0149707164
#> NIR40  -3.420457e-02  8.904731e-04 -0.067575181 -0.0129794353
#> NIR41  -3.265536e-02 -2.588608e-03 -0.068916396 -0.0073110346
#> NIR42  -2.984714e-02 -1.803429e-03 -0.066875767 -0.0058930953
#> NIR43  -2.941407e-02  1.944929e-03 -0.064192294 -0.0094845547
#> NIR44  -3.143770e-02  2.478976e-03 -0.065314906 -0.0123451695
#> NIR45  -3.351881e-02 -1.326440e-03 -0.069581823 -0.0108408723
#> NIR46  -3.428486e-02 -1.997008e-03 -0.072395354 -0.0120020038
#> NIR47  -3.366706e-02  5.099892e-03 -0.070835478 -0.0201267544
#> NIR48  -3.450930e-02  1.166034e-02 -0.068721263 -0.0287646505
#> NIR49  -3.838940e-02  1.221044e-02 -0.073091969 -0.0320795944
#> NIR50  -4.313153e-02  1.065212e-02 -0.081826871 -0.0342082384
#> NIR51  -4.565483e-02  4.180119e-03 -0.091100451 -0.0288728448
#> NIR52  -4.169298e-02 -1.720733e-02 -0.102892350  0.0033520365
#> NIR53  -3.224423e-02 -4.109102e-02 -0.111009318  0.0312488068
#> NIR54  -2.361946e-02 -4.916201e-02 -0.108675344  0.0452166366
#> NIR55  -1.654489e-02 -3.714856e-02 -0.096144163  0.0427179359
#> NIR56  -1.169685e-02 -1.126248e-02 -0.077475911  0.0110990005
#> NIR57  -1.014691e-02  1.205054e-02 -0.059610633 -0.0328958376
#> NIR58  -4.918889e-03  1.379837e-02 -0.053583705 -0.0340048642
#> NIR59   4.755685e-03  9.363066e-04 -0.051338427 -0.0050761772
#> NIR60   8.212777e-03 -1.132494e-02 -0.049392271  0.0310513967
#> NIR61   5.146947e-03 -2.045708e-02 -0.055677926  0.0616808751
#> NIR62   6.312422e-03 -2.203296e-02 -0.061318118  0.0749123900
#> NIR63   1.303421e-02 -1.611113e-02 -0.057347565  0.0731000696
#> NIR64   1.422116e-02 -1.106599e-02 -0.049990898  0.0622268747
#> NIR65   7.873790e-03 -5.394719e-03 -0.044993322  0.0356268993
#> NIR66   7.170677e-03  1.137927e-02 -0.044405978 -0.0022335031
#> NIR67   1.498220e-02  2.260739e-02 -0.037949291 -0.0196640746
#> NIR68   2.603544e-02  4.771798e-02 -0.042253436 -0.0362639904
#> NIR69   6.383787e-02  8.492879e-02 -0.029419805 -0.0440981976
#> NIR70   1.000377e-01  7.683935e-02  0.015756524  0.0322074785
#> NIR71   1.140048e-01  5.871186e-02  0.037834407  0.0656785598
#> NIR72   1.159581e-01  4.641634e-02  0.041084980  0.0804218607
#> NIR73   1.136673e-01  4.023444e-02  0.038877746  0.0870849094
#> NIR74   1.104668e-01  3.718318e-02  0.035795609  0.0885675815
#> NIR75   1.074530e-01  3.533359e-02  0.032300193  0.0878423805
#> NIR76   1.052701e-01  3.403376e-02  0.028759109  0.0863259121
#> NIR77   1.043708e-01  3.323780e-02  0.025196674  0.0849577160
#> NIR78   1.047859e-01  3.296461e-02  0.021159476  0.0839924398
#> NIR79   1.062758e-01  3.310066e-02  0.016048108  0.0838507526
#> NIR80   1.086711e-01  3.355627e-02  0.010651089  0.0832923225
#> NIR81   1.116293e-01  3.435968e-02  0.005122926  0.0814470373
#> NIR82   1.144054e-01  3.559394e-02 -0.002842356  0.0799478142
#> NIR83   1.162910e-01  3.718907e-02 -0.014626223  0.0784658950
#> NIR84   1.171046e-01  3.874999e-02 -0.030260015  0.0756533421
#> NIR85   1.168514e-01  3.970971e-02 -0.048030473  0.0728033513
#> NIR86   1.152502e-01  3.978679e-02 -0.067428446  0.0703528048
#> NIR87   1.128360e-01  3.919499e-02 -0.088907184  0.0658368362
#> NIR88   1.099585e-01  3.870950e-02 -0.109336951  0.0594282322
#> NIR89   1.061798e-01  3.918041e-02 -0.128356069  0.0570116722
#> NIR90   1.023012e-01  4.045748e-02 -0.144427214  0.0492969434
#> NIR91   9.972432e-02  4.182903e-02 -0.151108961  0.0310533687
#> NIR92   9.821457e-02  4.324393e-02 -0.153789402  0.0143836073
#> NIR93   9.695664e-02  4.495216e-02 -0.157315568  0.0018729507
#> NIR94   9.643888e-02  4.674486e-02 -0.160735779 -0.0047622602
#> NIR95   9.696313e-02  4.874639e-02 -0.166529740 -0.0056341168
#> NIR96   9.873766e-02  5.103590e-02 -0.169520543 -0.0034114349
#> NIR97   1.014125e-01  5.280361e-02 -0.168853034  0.0016049665
#> NIR98   1.036496e-01  5.364092e-02 -0.166881935  0.0066610016
#> NIR99   1.055234e-01  5.406348e-02 -0.158502231  0.0073974007
#> NIR100  1.076527e-01  5.431997e-02 -0.145161600  0.0053924823
#> NIR101  1.095304e-01  5.480715e-02 -0.133336801  0.0080037177
#> NIR102  1.104313e-01  5.550128e-02 -0.123542850  0.0140348117
#> NIR103  1.104275e-01  5.605672e-02 -0.114702394  0.0163248283
#> NIR104  1.101076e-01  5.672286e-02 -0.105386717  0.0174668109
#> NIR105  1.094454e-01  5.718261e-02 -0.094794469  0.0146807340
#> NIR106  1.081401e-01  5.747881e-02 -0.086857677  0.0090829702
#> NIR107  1.064748e-01  5.796668e-02 -0.082037612  0.0074065537
#> NIR108  1.046144e-01  5.847984e-02 -0.077351605  0.0054487896
#> NIR109  1.027069e-01  5.915674e-02 -0.072574421  0.0019643275
#> NIR110  1.012026e-01  6.010986e-02 -0.067458145 -0.0004461089
#> NIR111  9.991985e-02  6.119731e-02 -0.063152940 -0.0019100052
#> NIR112  9.855897e-02  6.247126e-02 -0.060846424 -0.0049773030
#> NIR113  9.720521e-02  6.404968e-02 -0.057631008 -0.0127890214
#> NIR114  9.578353e-02  6.602801e-02 -0.050312553 -0.0203587373
#> NIR115  9.407774e-02  6.845251e-02 -0.041758661 -0.0253913015
#> NIR116  9.260852e-02  7.099350e-02 -0.036711627 -0.0302888204
#> NIR117  9.194588e-02  7.370500e-02 -0.032396979 -0.0332279266
#> NIR118  9.171015e-02  7.667931e-02 -0.026681952 -0.0359895087
#> NIR119  9.138339e-02  7.944016e-02 -0.020316131 -0.0400471076
#> NIR120  9.086036e-02  8.186787e-02 -0.014125770 -0.0437547052
#> NIR121  9.024859e-02  8.407317e-02 -0.007268766 -0.0474658778
#> NIR122  8.948288e-02  8.595030e-02  0.000144654 -0.0492529617
#> NIR123  8.862589e-02  8.768225e-02  0.005843419 -0.0509385516
#> NIR124  8.799255e-02  8.993594e-02  0.010973947 -0.0535700717
#> NIR125  8.755197e-02  9.288096e-02  0.018132131 -0.0554048163
#> NIR126  8.712935e-02  9.559836e-02  0.026137437 -0.0571050401
#> NIR127  8.686731e-02  9.719827e-02  0.032801114 -0.0592570694
#> NIR128  8.624176e-02  9.769130e-02  0.040712505 -0.0588021837
#> NIR129  8.351762e-02  9.724070e-02  0.049869181 -0.0555330040
#> NIR130  7.827797e-02  9.584977e-02  0.055361227 -0.0537945825
#> NIR131  7.252987e-02  9.347109e-02  0.056763902 -0.0549473883
#> NIR132  6.777853e-02  8.983553e-02  0.056305098 -0.0556112304
#> NIR133  6.423042e-02  8.523301e-02  0.055025740 -0.0536194511
#> NIR134  6.144149e-02  8.065389e-02  0.053459576 -0.0513925219
#> NIR135  5.850488e-02  7.664893e-02  0.051546925 -0.0508126384
#> NIR136  5.469274e-02  7.274917e-02  0.049074120 -0.0510418495
#> NIR137  5.060985e-02  6.856976e-02  0.045762471 -0.0507534946
#> NIR138  4.764458e-02  6.440203e-02  0.042188012 -0.0497501262
#> NIR139  4.595081e-02  6.065337e-02  0.039452277 -0.0477682047
#> NIR140  4.455818e-02  5.675442e-02  0.037076397 -0.0442782448
#> NIR141  4.279654e-02  5.140861e-02  0.034715053 -0.0383298925
#> NIR142  3.997549e-02  4.374917e-02  0.033901300 -0.0301701687
#> NIR143  3.666373e-02  3.473691e-02  0.034526184 -0.0205316224
#> NIR144  3.619861e-02  2.754730e-02  0.032378853 -0.0123753158
#> NIR145  3.887940e-02  2.434911e-02  0.024815254 -0.0080509125
#> NIR146  3.964622e-02  2.407038e-02  0.016019080 -0.0064330623
#> NIR147  3.431430e-02  2.413091e-02  0.010130979 -0.0072798941
#> NIR148  2.297787e-02  2.241624e-02  0.004351753 -0.0116474971
#> NIR149  9.882733e-03  1.881219e-02 -0.004023871 -0.0175289455
#> NIR150  1.015366e-03  1.552048e-02 -0.012588764 -0.0215481617
#> NIR151 -5.956351e-04  1.412980e-02 -0.018059542 -0.0216838535
#> NIR152  3.095850e-03  1.275838e-02 -0.019839973 -0.0148010848
#> NIR153  7.383891e-03  1.065651e-02 -0.019763003 -0.0046064203
#> NIR154  9.604775e-03  1.023867e-02 -0.019405335  0.0012561331
#> NIR155  1.168321e-02  1.118619e-02 -0.018497799  0.0037248946
#> NIR156  1.538586e-02  1.114153e-02 -0.016208243  0.0080697437
#> NIR157  1.911424e-02  1.089016e-02 -0.012762796  0.0123855983
#> NIR158  2.072069e-02  1.197687e-02 -0.009766196  0.0131897788
#> NIR159  2.008720e-02  1.366036e-02 -0.008488977  0.0109448112
#> NIR160  1.800666e-02  1.477616e-02 -0.009105847  0.0080877299
#> NIR161  1.590361e-02  1.499982e-02 -0.010852986  0.0061108417
#> NIR162  1.607253e-02  1.413900e-02 -0.011583042  0.0069214812
#> NIR163  1.760846e-02  1.251786e-02 -0.009920632  0.0104142298
#> NIR164  1.698938e-02  1.101840e-02 -0.009082676  0.0124530089
#> NIR165  1.411756e-02  9.761127e-03 -0.011160606  0.0117093687
#> NIR166  1.108416e-02  8.574941e-03 -0.014399733  0.0103945249
#> NIR167  8.793373e-03  7.495023e-03 -0.017243291  0.0098252013
#> NIR168  7.042095e-03  6.571564e-03 -0.019282794  0.0095952645
#> NIR169  5.589539e-03  5.767249e-03 -0.020861677  0.0094821855
#> NIR170  4.362690e-03  5.047961e-03 -0.022367411  0.0096135325
#> NIR171  3.323070e-03  4.410634e-03 -0.023676003  0.0097059875
#> NIR172  2.462306e-03  3.866898e-03 -0.024689108  0.0096672744
#> NIR173  1.785351e-03  3.423101e-03 -0.025580229  0.0097140099
#> NIR174  1.229319e-03  3.044163e-03 -0.026357191  0.0097194158
#> NIR175  7.251181e-04  2.689458e-03 -0.026894566  0.0096281260
#> NIR176  2.650787e-04  2.363551e-03 -0.027424673  0.0096757240
#> NIR177 -1.696646e-04  2.078286e-03 -0.027962880  0.0097378822
#> NIR178 -6.058720e-04  1.818866e-03 -0.028371513  0.0095941394
#> NIR179 -1.009434e-03  1.583418e-03 -0.028781580  0.0094562276
#> NIR180 -1.351203e-03  1.389738e-03 -0.029243834  0.0094683509
#> NIR181 -1.644406e-03  1.245508e-03 -0.029568328  0.0093058851
#> NIR182 -1.870415e-03  1.147624e-03 -0.029777394  0.0090888707
#> NIR183 -2.000738e-03  1.081875e-03 -0.030059637  0.0091075219
#> NIR184 -2.080340e-03  9.915259e-04 -0.030216244  0.0091819066
#> NIR185 -2.155584e-03  8.521834e-04 -0.030236784  0.0092292950
#> NIR186 -2.217807e-03  7.121934e-04 -0.030359661  0.0093970598
#> NIR187 -2.271436e-03  6.259727e-04 -0.030491980  0.0094767326
#> NIR188 -2.321780e-03  5.882508e-04 -0.030528943  0.0093607609
#> NIR189 -2.317601e-03  5.695086e-04 -0.030574769  0.0093732713
#> NIR190 -2.240825e-03  5.383054e-04 -0.030639885  0.0095234849
#> NIR191 -2.129850e-03  4.697268e-04 -0.030518135  0.0096510729
#> NIR192 -1.992368e-03  3.587620e-04 -0.030381115  0.0098637546
#> NIR193 -1.827479e-03  2.327608e-04 -0.030369920  0.0102537832
#> NIR194 -1.647574e-03  1.446636e-04 -0.030261211  0.0104321921
#> NIR195 -1.424722e-03  1.558471e-04 -0.030055638  0.0103939954
#> NIR196 -1.145519e-03  2.691782e-04 -0.029879279  0.0104697444
#> NIR197 -8.754454e-04  3.971566e-04 -0.029633880  0.0105529140
#> NIR198 -7.013720e-04  4.800008e-04 -0.029276021  0.0105335209
#> NIR199 -7.346804e-04  5.226082e-04 -0.029065195  0.0105229560
#> NIR200 -1.144657e-03  4.958769e-04 -0.028995498  0.0103201202
#> NIR201 -1.994776e-03  3.441844e-04 -0.029071118  0.0097550282
#> NIR202 -3.181979e-03  8.727982e-05 -0.029537913  0.0090828959
#> NIR203 -4.586710e-03 -2.194379e-04 -0.030380408  0.0083440983
#> NIR204 -6.111049e-03 -5.184947e-04 -0.031356285  0.0071941252
#> NIR205 -7.624620e-03 -7.448019e-04 -0.032425929  0.0058920654
#> NIR206 -9.035376e-03 -9.386728e-04 -0.033623440  0.0047630819
#> NIR207 -1.036141e-02 -1.178173e-03 -0.034712018  0.0036843593
#> NIR208 -1.164530e-02 -1.413282e-03 -0.035668366  0.0025415421
#> NIR209 -1.289320e-02 -1.558695e-03 -0.036687612  0.0013965988
#> NIR210 -1.400392e-02 -1.635169e-03 -0.037825280  0.0001961030
#> NIR211 -1.471900e-02 -1.741658e-03 -0.039013056 -0.0006973855
#> NIR212 -1.489531e-02 -2.010503e-03 -0.040143930 -0.0007412380
#> NIR213 -1.471450e-02 -2.571973e-03 -0.040985598 -0.0001860462
#> NIR214 -1.448544e-02 -3.372924e-03 -0.041443918  0.0005043870
#> NIR215 -1.434284e-02 -4.250369e-03 -0.041926353  0.0013275375
#> NIR216 -1.427168e-02 -5.208370e-03 -0.042562107  0.0022344402
#> NIR217 -1.427435e-02 -6.478388e-03 -0.043224660  0.0033443663
#> NIR218 -1.437416e-02 -8.449011e-03 -0.044257841  0.0052085089
#> NIR219 -1.429078e-02 -1.134277e-02 -0.046049610  0.0084169469
#> NIR220 -1.316021e-02 -1.475193e-02 -0.048279340  0.0133349560
#> NIR221 -1.014430e-02 -1.785284e-02 -0.049770954  0.0196534220
#> NIR222 -5.677561e-03 -2.007768e-02 -0.049051965  0.0261994711
#> NIR223 -1.724310e-03 -2.138696e-02 -0.045686082  0.0313832588
#> NIR224 -8.615735e-05 -2.200911e-02 -0.041991127  0.0341138331
#> NIR225 -7.016698e-04 -2.200280e-02 -0.040335467  0.0346387987
#> NIR226 -2.808897e-03 -2.107052e-02 -0.040016773  0.0328013645
#> NIR227 -6.665507e-03 -1.830327e-02 -0.040100315  0.0269896462
#> NIR228 -1.236831e-02 -1.283891e-02 -0.041441445  0.0159587846
#> NIR229 -1.811204e-02 -5.622938e-03 -0.044088354  0.0018283297
#> NIR230 -2.095699e-02  9.552249e-04 -0.046087162 -0.0099942430
#> NIR231 -1.926906e-02  5.009288e-03 -0.045590012 -0.0147813697
#> NIR232 -1.420356e-02  6.048327e-03 -0.042251564 -0.0118345542
#> NIR233 -8.635851e-03  5.419253e-03 -0.037472176 -0.0057014629
#> NIR234 -4.862262e-03  4.728888e-03 -0.033061831 -0.0006465373
#> NIR235 -3.387811e-03  4.242864e-03 -0.030245509  0.0024738679
#> NIR236 -3.141614e-03  3.798996e-03 -0.029318866  0.0039181690
#> NIR237 -2.763360e-03  3.532957e-03 -0.029369914  0.0044793820
#> NIR238 -1.843574e-03  3.447565e-03 -0.029349372  0.0051375454
#> NIR239 -7.314476e-04  3.293330e-03 -0.028862743  0.0060830808
#> NIR240  3.511386e-04  2.767237e-03 -0.028215872  0.0075911227
#> NIR241  1.377964e-03  1.786167e-03 -0.027842190  0.0099476563
#> NIR242  2.336498e-03  7.202884e-04 -0.027699706  0.0124419365
#> NIR243  3.274575e-03 -7.117817e-05 -0.027390616  0.0143006269
#> NIR244  3.937034e-03 -6.891943e-04 -0.026814726  0.0157360414
#> NIR245  3.783073e-03 -1.240316e-03 -0.026413082  0.0165795870
#> NIR246  2.655571e-03 -1.448487e-03 -0.026445516  0.0161874811
#> NIR247  9.107471e-04 -1.052198e-03 -0.026943678  0.0144483729
#> NIR248 -1.020941e-03 -1.569609e-04 -0.027834756  0.0117750588
#> NIR249 -2.485476e-03  1.027827e-03 -0.028878924  0.0087910074
#> NIR250 -2.784939e-03  2.162468e-03 -0.029428840  0.0068429354
#> NIR251 -2.192507e-03  2.797800e-03 -0.029143282  0.0067483235
#> NIR252 -1.550832e-03  2.820878e-03 -0.028441419  0.0075937313
#> NIR253 -1.150048e-03  2.372312e-03 -0.027874761  0.0085097467
#> NIR254 -8.803887e-04  1.669410e-03 -0.027735229  0.0095857821
#> NIR255 -7.700849e-04  9.317159e-04 -0.027862985  0.0107138382
#> NIR256 -8.602223e-04  2.969347e-04 -0.028117601  0.0115600888
#> NIR257 -1.082589e-03 -2.006354e-04 -0.028629032  0.0122671783
#> NIR258 -1.413385e-03 -5.893277e-04 -0.029238941  0.0126753723
#> NIR259 -1.828582e-03 -8.688813e-04 -0.029746279  0.0125943651
#> NIR260 -2.226934e-03 -1.009854e-03 -0.030213772  0.0123566585
#> NIR261 -2.524898e-03 -1.092762e-03 -0.030681112  0.0122204618
#> NIR262 -2.749303e-03 -1.222047e-03 -0.030967404  0.0121156781
#> NIR263 -2.931166e-03 -1.353408e-03 -0.031168325  0.0120799644
#> NIR264 -3.061128e-03 -1.415451e-03 -0.031435724  0.0121735783
#> NIR265 -3.160593e-03 -1.429246e-03 -0.031593853  0.0121906510
#> NIR266 -3.227588e-03 -1.441294e-03 -0.031611122  0.0120667974
#> NIR267 -3.252319e-03 -1.467328e-03 -0.031650126  0.0120601568
#> NIR268 -3.255688e-03 -1.508291e-03 -0.031718130  0.0121695820